% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{addQuotes}
\alias{addQuotes}
\title{Add double quotes to a string or to each element of a string vector}
\usage{
addQuotes(txt)
}
\arguments{
\item{txt}{Character vector}
}
\value{
The double-quoted character vector
}
\description{
Add double quotes to a string or to each element of a string vector
}
\keyword{internal}
