% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{blocDiag}
\alias{blocDiag}
\title{Bloc-diagonal matrix constructor}
\usage{
blocDiag(blocs)
}
\arguments{
\item{blocs}{list, each element is a square matrix.}
}
\value{
A square matrix.
}
\description{
This function creates a square bloc-diagonal matrix from a list of square blocs.
}
\examples{
blocDiag(list(1,cbind(c(1,2),c(3,4)),25))
}
