% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{checkFormula}
\alias{checkFormula}
\title{Check a formula make sense within a given namespace}
\usage{
checkFormula(f, namespace)
}
\arguments{
\item{f}{Character string, formula}

\item{namespace}{Character vector, namespace (i.e. names of known variables)}
}
\value{
A list with the following fields:
\describe{
  \item{ok}{Does the formula make sense?}
  \item{mess}{message}
}
}
\description{
Check a formula make sense within a given namespace
}
\keyword{internal}
