% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{densityPlot}
\alias{densityPlot}
\title{densityPlot}
\usage{
densityPlot(sim, xlab = "values", col = "black")
}
\arguments{
\item{sim}{vector or matrix or data frame, MCMC simulations}

\item{xlab}{Character, label of x-axis to be used if sim has no names}

\item{col}{Color}
}
\value{
A ggplot (or a list thereof if several columns in sim)
}
\description{
returns a histogram+density ggplot (or a list thereof if several columns in sim)
}
