% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{foundBaM}
\alias{foundBaM}
\title{Was BaM executable found?}
\usage{
foundBaM(exedir, exename)
}
\arguments{
\item{exedir}{directory where the executable should be searched}

\item{exename}{name of the executable}
}
\value{
a logical, TRUE if the exe was found, false otherwise
}
\description{
Was BaM executable found?
}
\keyword{internal}
