% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI_modelSpecific.R
\name{getCatalogue}
\alias{getCatalogue}
\title{BaM catalogue}
\usage{
getCatalogue(printOnly = FALSE)
}
\arguments{
\item{printOnly}{Logical, should the catalogue be returned or only printed?}
}
\value{
If \code{printOnly==FALSE}, a list with the following fields:
\describe{
  \item{distributions}{available univariate distributions.}
  \item{models}{available models.}
}
}
\description{
Distributions and models available in BaM
}
\examples{
catalogue <- getCatalogue()
getCatalogue(printOnly=TRUE)
}
