% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_parameter.R
\name{getInitPar}
\alias{getInitPar}
\title{Get initial values}
\usage{
getInitPar(parameters)
}
\arguments{
\item{parameters}{List of parameter objects, parameters whose initial values are sought.}
}
\value{
A numeric vector containing the initial values
}
\description{
Get the initial values of a list of parameters and return them as a vector.
}
\examples{
ps <- list(parameter(name='par1',init=0),parameter(name='par2',init=1))
getInitPar(ps)
}
