% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getNames}
\alias{getNames}
\title{Get object names}
\usage{
getNames(loo, name = "name")
}
\arguments{
\item{loo}{List Of Objects}

\item{name}{character, string denoting the name field}
}
\value{
A character vector containing names
}
\description{
getNames from an object or a list of objects having a $name field (e.g. parameters)
}
\examples{
pars <- list(parameter(name='par1',init=0),
             parameter(name='par2',init=0),
             parameter(name='Third parameter',init=0))
getNames(pars)
}
