% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getParNames}
\alias{getParNames}
\title{Get parameter names}
\usage{
getParNames(d)
}
\arguments{
\item{d}{Character (possibly vector), distribution (possibly distributions)}
}
\value{
A character vector with parameter names.
}
\description{
Get parameter names for a distribution d
}
\examples{
parnames <- getParNames('GEV')
npar <- length(getParNames('Gumbel'))
}
