% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{getPathToBaM}
\alias{getPathToBaM}
\title{Path to BaM}
\usage{
getPathToBaM()
}
\value{
A string containing the path of the directory containing BaM executable,
or NULL if this directory has not been set yet.
}
\description{
Get path to BaM executable
}
\examples{
  getPathToBaM()
}
