% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dataset.R
\name{is.dataset}
\alias{is.dataset}
\title{dataset tester}
\usage{
is.dataset(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'dataset', FALSE otherwise.
}
\description{
Is an object of class 'dataset'?
}
\keyword{internal}
