% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcCooking.R
\name{is.mcmcCooking}
\alias{is.mcmcCooking}
\title{mcmcCooking tester}
\usage{
is.mcmcCooking(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'mcmcCooking', FALSE otherwise.
}
\description{
Is an object of class 'mcmcCooking'?
}
\keyword{internal}
