% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcOptions.R
\name{is.mcmcOptions}
\alias{is.mcmcOptions}
\title{mcmcOptions tester}
\usage{
is.mcmcOptions(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'mcmcOptions', FALSE otherwise.
}
\description{
Is an object of class 'mcmcOptions'?
}
\keyword{internal}
