% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcSummary.R
\name{is.mcmcSummary}
\alias{is.mcmcSummary}
\title{mcmcSummary tester}
\usage{
is.mcmcSummary(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'mcmcSummary', FALSE otherwise.
}
\description{
Is an object of class 'mcmcSummary'?
}
\keyword{internal}
