% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{is.model}
\alias{is.model}
\title{model tester}
\usage{
is.model(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'model', FALSE otherwise.
}
\description{
Is an object of class 'model'?
}
\keyword{internal}
