% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_parameter_VAR.R
\name{is.parameter_VAR}
\alias{is.parameter_VAR}
\title{parameter_VAR tester}
\usage{
is.parameter_VAR(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'parameter_VAR', FALSE otherwise.
}
\description{
Is an object of class 'parameter_VAR'?
}
\keyword{internal}
