% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prediction.R
\name{is.prediction}
\alias{is.prediction}
\title{prediction tester}
\usage{
is.prediction(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'prediction', FALSE otherwise.
}
\description{
Is an object of class 'prediction'?
}
\keyword{internal}
