% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_remnantErrorModel.R
\name{is.remnantErrorModel}
\alias{is.remnantErrorModel}
\title{remnantErrorModel tester}
\usage{
is.remnantErrorModel(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'remnantErrorModel', FALSE otherwise.
}
\description{
Is an object of class 'remnantErrorModel'?
}
\keyword{internal}
