% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_residualOptions.R
\name{is.residualOptions}
\alias{is.residualOptions}
\title{residualOptions tester}
\usage{
is.residualOptions(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'residualOptions', FALSE otherwise.
}
\description{
Is an object of class 'residualOptions'?
}
\keyword{internal}
