% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_runOptions.R
\name{is.runOptions}
\alias{is.runOptions}
\title{runOptions tester}
\usage{
is.runOptions(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'runOptions', FALSE otherwise.
}
\description{
Is an object of class 'runOptions'?
}
\keyword{internal}
