% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_xtraModelInfo.R
\name{is.xtraModelInfo}
\alias{is.xtraModelInfo}
\title{xtraModelInfo tester}
\usage{
is.xtraModelInfo(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'xtraModelInfo', FALSE otherwise.
}
\description{
Is an object of class 'xtraModelInfo'?
}
\keyword{internal}
