% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceFunk.R
\name{llfunk_iid_Gaussian}
\alias{llfunk_iid_Gaussian}
\title{Log-likelihood function: iid Gaussian}
\usage{
llfunk_iid_Gaussian(Ysim, Yobs, Yu, gamma)
}
\arguments{
\item{Ysim}{data frame, model-simulated values.}

\item{Yobs}{data frame, corresponding observed values, same dimensions as Ysim. NAs are skipped.}

\item{Yu}{data frame, measurement uncertainties (standard deviations), same dimensions as Ysim and Yobs.}

\item{gamma}{numeric vector, structural error parameters. length(gamma) = number of columns in Ysim.}
}
\value{
A numeric value equal to the log-likelihood.
}
\description{
Computes the log-likelihood from model-simulated values, based on a Gaussian iid error model:
\itemize{
    \item Yobs = Ysim + ğ›¿ + ğœ€
    \item Measurement errors: ğ›¿~ N(0,sdev=Yu)
    \item Structural errors: ğœ€~ N(0,sdev=ğ›¾)
}
If Yobs/Ysim are multi-variate, this error model is applied independently to each component.
}
\examples{
Yobs=SauzeGaugings['Q']
Yu=SauzeGaugings['uQ']
Ysim=100*(SauzeGaugings['H']+0.5)^1.6
llfunk_iid_Gaussian(Ysim,Yobs,Yu,gamma=100)
}
