% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferenceFunk.R
\name{logPrior_Flat}
\alias{logPrior_Flat}
\title{Log-prior function: improper flat prior}
\usage{
logPrior_Flat(parvector)
}
\arguments{
\item{parvector}{numeric vector, parameter vector, including thetas (model parameters)
and gammas (structural errors parameters).}
}
\value{
A numeric value equal to the prior log-density.
}
\description{
Computes the log-density of an improper flat prior distribution.
}
\examples{
logPrior_Flat(c(1,1,0.2))
}
