% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcCooking.R
\name{mcmcCooking}
\alias{mcmcCooking}
\title{mcmcCooking constructor.}
\usage{
mcmcCooking(
  fname = "Config_Cooking.txt",
  result.fname = "Results_Cooking.txt",
  burn = 0.5,
  nSlim = 10
)
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{result.fname}{Character, result file name.}

\item{burn}{numeric, burn factor, >=0 and <1. 0.4 means the first 40% of
MCMC samples are discarded).}

\item{nSlim}{Integer, slimming period: 10 means only one MCMC sample
every 10 is kept (after burning).}
}
\value{
An object of class 'mcmcCooking'.
}
\description{
Creates a new instance of a 'mcmcCooking' object
}
\examples{
m <- mcmcCooking()
}
