% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcSummary.R
\name{mcmcSummary}
\alias{mcmcSummary}
\title{mcmcSummary constructor.}
\usage{
mcmcSummary(
  fname = "Config_Summary.txt",
  result.fname = "Results_Summary.txt",
  DIC.fname = "Results_DIC.txt",
  xtendedMCMC.fname = ""
)
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{result.fname}{Character, summary file name.}

\item{DIC.fname}{Character, DIC file name. Not computed if empty string.}

\item{xtendedMCMC.fname}{Character, xtended MCMC file name. Not written if empty string.}
}
\value{
An object of class 'mcmcSummary'.
}
\description{
Creates a new instance of a 'mcmcSummary' object
}
\examples{
m <- mcmcSummary()
}
