% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MeyrasGaugings}
\alias{MeyrasGaugings}
\title{Meyras Gaugings}
\format{
A data frame with 104 rows and 4 variables:
\describe{
  \item{h}{Stage (m)}
  \item{Q}{Discharge (m3/s)}
  \item{uQ}{Discharge uncertainty (m3/s) expressed as a standard deviation}
  \item{Period}{Stability period on which a single rating curve can be used}
}
}
\usage{
MeyrasGaugings
}
\description{
Stage-discharge gaugings from the hydrometric station 'the Ardèche River at Meyras'.
See https://en.wikipedia.org/wiki/Ardèche_(river) for a description of the river
See https://doi.org/10.1029/2018WR023389 for an article using this dataset
}
\keyword{datasets}
