% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_parameter.R
\name{parameter}
\alias{parameter}
\title{parameter object constructor.}
\usage{
parameter(name, init, prior.dist = "FlatPrior", prior.par = NULL)
}
\arguments{
\item{name}{character, parameter name.}

\item{init}{numeric, initial guess.}

\item{prior.dist}{character, prior distribution.}

\item{prior.par}{numeric vector, prior parameters}
}
\value{
An object of class 'parameter'.
}
\description{
Creates a new instance of a 'parameter' object
}
\examples{
p <- parameter(name='par',init=0,prior.dist='Gaussian',prior.par=c(0,1))
}
