% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_remnantErrorModel.R
\name{remnantErrorModel}
\alias{remnantErrorModel}
\title{remnantErrorModel object constructor.}
\usage{
remnantErrorModel(
  fname = "Config_RemnantSigma.txt",
  funk = "Linear",
  par = list(parameter("g1", 1, prior.dist = "FlatPrior+"), parameter("g2", 0.1,
    prior.dist = "FlatPrior+"))
)
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{funk}{Character, function f used in remnant sdev = f(Ysim).
Available: 'Constant','Linear' (default), 'Exponential', 'Gaussian'.}

\item{par}{list of parameter objects, parameters of the function above.
respectively, npar= 1,2,3,3}
}
\value{
An object of class 'remnantErrorModel'.
}
\description{
Creates a new instance of a 'remnantErrorModel' object
}
\examples{
r <- remnantErrorModel()
}
