% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_residualOptions.R
\name{residualOptions}
\alias{residualOptions}
\title{residualOptions constructor.}
\usage{
residualOptions(
  fname = "Config_Residuals.txt",
  result.fname = "Results_Residuals.txt"
)
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{result.fname}{Character, result file name.}
}
\value{
An object of class 'residualOptions'.
}
\description{
Creates a new instance of a 'residualOptions' object
}
\examples{
r <- residualOptions()
}
