% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{runExe}
\alias{runExe}
\title{Run BaM executable}
\usage{
runExe(exedir, exename, workspace, arguments = NULL, stout = "")
}
\arguments{
\item{exedir}{Character string, directory of the executable}

\item{exename}{Character string, name of the executable}

\item{workspace}{Character string, full path to workspace}

\item{arguments}{Character string, arguments to be passed to BaM exe}

\item{stout}{Character string, standard output (see ?system2).
In particular, stout="" (default) shows BaM messages in the console,
stout=NULL discards BaM messages,
stout='log.txt' saves BaM messages in file "log.txt".}
}
\value{
an error code (0 for success)
}
\description{
Run BaM executable
}
\keyword{internal}
