% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_runOptions.R
\name{runOptions}
\alias{runOptions}
\title{runOptions constructor.}
\usage{
runOptions(
  fname = "Config_RunOptions.txt",
  doMCMC = TRUE,
  doSummary = TRUE,
  doResiduals = TRUE,
  doPrediction = FALSE
)
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{doMCMC}{logical, do MCMC sampling?}

\item{doSummary}{logical, do MCMC summarizing?}

\item{doResiduals}{logical, do residuals analysis?}

\item{doPrediction}{logical, do prediction experiments?}
}
\value{
An object of class 'runOptions'.
}
\description{
Creates a new instance of a 'runOptions' object
}
\examples{
o <- runOptions()
}
