% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SauzeGaugings}
\alias{SauzeGaugings}
\title{Sauze Gaugings}
\format{
A data frame with 38 rows and 3 variables:
\describe{
  \item{H}{Stage (m)}
  \item{Q}{Discharge (m3/s)}
  \item{uQ}{Discharge uncertainty (m3/s) expressed as a standard deviation}
}
}
\usage{
SauzeGaugings
}
\description{
Stage-discharge gaugings from the hydrometric station 'the Ardèche River at Sauze-St-Martin'.
See https://en.wikipedia.org/wiki/Ardèche_(river) for a description of the river
See https://hal.science/hal-00934237 for an article using this dataset
}
\keyword{datasets}
