% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_parameter.R
\name{setInitPar}
\alias{setInitPar}
\title{Set initial values}
\usage{
setInitPar(parameters, values)
}
\arguments{
\item{parameters}{List of parameter objects, parameters whose initial values are to be set.}

\item{values}{Numeric vector, initial values.}
}
\value{
a list of parameter objects equal to the input list 'parameters', except
    for their initial values.
}
\description{
Set the initial values of a list of parameters.
}
\examples{
ps <- list(parameter(name='par1',init=0),parameter(name='par2',init=1))
ps=setInitPar(ps,c(10,100))
print(ps)
}
