% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{setPathToBaM}
\alias{setPathToBaM}
\title{Path to BaM}
\usage{
setPathToBaM(dir.exe, quiet = FALSE)
}
\arguments{
\item{dir.exe}{character string, folder where BaM executable is located.}

\item{quiet}{logical, if TRUE, suppress status messages.}
}
\value{
nothing - just write a config file.
}
\description{
Set path to BaM executable
}
\examples{
  setPathToBaM(dir.exe=tempdir())
}
