% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dataset.R
\name{toString.dataset}
\alias{toString.dataset}
\title{dataset to string}
\usage{
\method{toString}{dataset}(x, ...)
}
\arguments{
\item{x}{dataset object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'dataset' into a ready-to-write vector of string
}
\examples{
X=data.frame(input1=rnorm(100),input2=rnorm(100))
Y=data.frame(output=X$input1+0.8*X$input2+0.1*rnorm(100))
workspace=tempdir()
d <- dataset(X=X,Y=Y,data.dir=workspace)
toString(d)
}
