% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcCooking.R
\name{toString.mcmcCooking}
\alias{toString.mcmcCooking}
\title{mcmcCooking to string}
\usage{
\method{toString}{mcmcCooking}(x, ...)
}
\arguments{
\item{x}{mcmcCooking object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'mcmcCooking' into a ready-to-write vector of string
}
\examples{
toString(mcmcCooking())
}
