% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcOptions.R
\name{toString.mcmcOptions}
\alias{toString.mcmcOptions}
\title{mcmcOptions to string}
\usage{
\method{toString}{mcmcOptions}(x, ...)
}
\arguments{
\item{x}{mcmcOptions object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'mcmcOptions' into a ready-to-write vector of string
}
\examples{
toString(mcmcOptions())
}
