% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcSummary.R
\name{toString.mcmcSummary}
\alias{toString.mcmcSummary}
\title{mcmcSummary to string}
\usage{
\method{toString}{mcmcSummary}(x, ...)
}
\arguments{
\item{x}{mcmcSummary object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'mcmcSummary' into a ready-to-write vector of string
}
\examples{
toString(mcmcSummary())
}
