% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{toString.model}
\alias{toString.model}
\title{model to string}
\usage{
\method{toString}{model}(x, ...)
}
\arguments{
\item{x}{model object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'model' into a ready-to-write vector of string
}
\examples{
toString(model())
}
