% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_parameter.R
\name{toString.parameter}
\alias{toString.parameter}
\title{parameter to string}
\usage{
\method{toString}{parameter}(x, ...)
}
\arguments{
\item{x}{parameter object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'parameter' into a ready-to-write vector of string
}
\examples{
p <- parameter(name='par',init=0,prior.dist='Gaussian',prior.par=c(0,1))
toString(p)
}
