% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prediction.R
\name{toString.prediction}
\alias{toString.prediction}
\title{prediction to string}
\usage{
\method{toString}{prediction}(x, ...)
}
\arguments{
\item{x}{prediction object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'prediction' into a ready-to-write vector of string
}
\examples{
pred=prediction(X=twoPopulations[,1:3],spagFiles=c('P1.spag','P2.spag'))
toString(pred)
}
