% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_remnantErrorModel.R
\name{toString.remnantErrorModel}
\alias{toString.remnantErrorModel}
\title{remnantErrorModel to string}
\usage{
\method{toString}{remnantErrorModel}(x, ...)
}
\arguments{
\item{x}{remnantErrorModel object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'remnantErrorModel' into a ready-to-write vector of string
}
\examples{
toString(remnantErrorModel())
}
