% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_residualOptions.R
\name{toString.residualOptions}
\alias{toString.residualOptions}
\title{residualOptions to string}
\usage{
\method{toString}{residualOptions}(x, ...)
}
\arguments{
\item{x}{residualOptions object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'residualOptions' into a ready-to-write vector of string
}
\examples{
toString(residualOptions())
}
