% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_runOptions.R
\name{toString.runOptions}
\alias{toString.runOptions}
\title{runOptions to string}
\usage{
\method{toString}{runOptions}(x, ...)
}
\arguments{
\item{x}{runOptions object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'runOptions' into a ready-to-write vector of string
}
\examples{
toString(runOptions())
}
