% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toString_engine}
\alias{toString_engine}
\title{Write value-comment pairs into string vector}
\usage{
toString_engine(val, comment, addQuote = T)
}
\arguments{
\item{val}{List, object/list of values to be printed}

\item{comment}{Character vector, associated comments (same length as val)}

\item{addQuote}{Logical, ass double quotes to strings?}
}
\value{
A character vector, ready to be printed of written to file
}
\description{
Write value-comment pairs into string vector
}
\keyword{internal}
