% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{tracePlot}
\alias{tracePlot}
\title{MCMC trace plots}
\usage{
tracePlot(sim, ylab = "values", keep = NULL, col = "black", psize = 0.5)
}
\arguments{
\item{sim}{vector or matrix or data frame, MCMC simulations}

\item{ylab}{Character, label of y-axis to be used if sim has no names}

\item{keep}{Integer vector, indices of samples to be kept in cooked MCMC sample}

\item{col}{Color}

\item{psize}{Numeric, point size}
}
\value{
A ggplot (or a list thereof if several columns in sim)
}
\description{
2DO (adapt from STooDs): Generate a trace of MCMC samples for all variables
MCMC parameter plots
}
\details{
2DO (adapt from STooDs): MCMC-based estimation of marginal posterior pdf for each parameter
MCMC reporting

2DO (adapt from STooDs): Generate pdf report files summarizing mcmc samples

tracePlot

returns a trace plot ggplot (or a list thereof if several columns in sim)
}
