% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoPopulations}
\alias{twoPopulations}
\title{Evolution of two populations}
\format{
An object of class \code{data.frame} with 101 rows and 5 columns.
}
\usage{
twoPopulations
}
\description{
Size of two populations of the same species put in two different environments,
as a function of time and local temperature.
\itemize{
  \item Input variables: time t, temperature at site 1 T1, temperature at site 2 T2.
  \item Output variables: population size at site 1 P1, population size at site 2 P2.
  \item Data are synthetically generated from a logistic model.
}
}
\keyword{datasets}
