% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{violinPlot}
\alias{violinPlot}
\title{violinPlot}
\usage{
violinPlot(sim, ylab = "values", col = "black")
}
\arguments{
\item{sim}{vector or matrix or data frame, MCMC simulations}

\item{ylab}{Character, label of y-axis}

\item{col}{Color}
}
\value{
A ggplot
}
\description{
returns a violinplot ggplot
}
