% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI_modelSpecific.R
\name{writeConfig.xtra}
\alias{writeConfig.xtra}
\title{Xtra Model Information}
\usage{
writeConfig.xtra(workspace, mod)
}
\arguments{
\item{workspace}{Character, directory where config and result files are stored.}

\item{mod}{model object, the model to be calibrated}
}
\value{
Nothing: just write config files
}
\description{
Handle Xtra Model Information
In particular, write Model Xtra configuration file, but could also run a
script to pre-load data, format stuff, etc.
}
