% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_prediction.R
\name{writePredInputs}
\alias{writePredInputs}
\title{Write prediction inputs}
\usage{
writePredInputs(o)
}
\arguments{
\item{o}{prediction object}
}
\value{
nothing - just write to files.
}
\description{
Write input data of the prediction into files
}
\examples{
temp=tempdir()
pred=prediction(X=twoPopulations[,1:3],spagFiles=c('P1.spag','P2.spag'),
                data.dir=temp)
writePredInputs(pred)
}
