% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_xtraModelInfo.R
\name{xtraModelInfo}
\alias{xtraModelInfo}
\title{xtraModelInfo constructor.}
\usage{
xtraModelInfo(fname = "Config_Xtra.txt", object = NULL)
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{object}{any R object containing xtra model info - typically a list of stuff.
The content and meaning of 'object' is completely model-specific.}
}
\value{
An object of class 'xtraModelInfo'.
}
\description{
Creates a new instance of a 'xtraModelInfo' object containing extra model information
}
\examples{
x <- xtraModelInfo()
}
