% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{BaM}
\alias{BaM}
\title{Run BaM}
\usage{
BaM(
  mod,
  data,
  remnant = rep(list(remnantErrorModel()), mod$nY),
  mcmc = mcmcOptions(),
  cook = mcmcCooking(),
  summary = mcmcSummary(),
  residuals = residualOptions(),
  pred = NULL,
  doCalib = TRUE,
  doPred = FALSE,
  na.value = -9999,
  run = TRUE,
  preClean = FALSE,
  workspace = file.path(getwd(), "BaM_workspace"),
  dir.exe = file.path(path.package("RBaM"), "bin"),
  name.exe = "BaM",
  predMaster_fname = "Config_Pred_Master.txt"
)
}
\arguments{
\item{mod}{model object, the model to be calibrated}

\item{data}{dataset object, calibration data}

\item{remnant}{list of remnantErrorModel objects.
WARNING: make sure you use a list of length mod$nY (even if mod$nY=1!)}

\item{mcmc}{mcmcOptions object, MCMC simulation number and options}

\item{cook}{mcmcCooking object, properties of MCMC cooking (burn and slice)}

\item{summary}{mcmcSummary object, properties of MCMC summary}

\item{residuals}{residualOptions object, properties of residual analysis}

\item{pred}{list of prediction objects, properties of prediction experiments}

\item{doCalib}{Logical, do Calibration? (mcmc+cooking+summary+residuals)}

\item{doPred}{Logical, do Prediction?}

\item{na.value}{numeric, value used for NAs when writing the dataset in BaM format}

\item{run}{Logical, run BaM? if FALSE, just write config files.}

\item{preClean}{Logical, start by cleaning up workspace?
Be careful, this will delete all files in the workspace, including old results!}

\item{workspace}{Character, directory where config and result files are stored.}

\item{dir.exe}{Character, directory where BaM executable stands.}

\item{name.exe}{Character, name of the executable without extension ('BaM' by default).}

\item{predMaster_fname}{Character, name of configuration file pointing to all prediction experiments.}
}
\value{
Nothing: just write config files and runs the executable.
}
\description{
Run BaM.exe
}
