% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{densityPlot}
\alias{densityPlot}
\title{densityPlot}
\usage{
densityPlot(sim, xlab = "values", col = "black")
}
\arguments{
\item{sim}{vector or matrix or data frame, MCMC simulations}

\item{xlab}{Character, label of x-axis to be used if sim has no names}

\item{col}{Color}
}
\value{
A ggplot (or a list thereof if several columns in sim)
}
\description{
returns a histogram+density ggplot (or a list thereof if several columns in sim)
}
\examples{
# Create Monte Carlo samples
n=1000
sim=data.frame(p1=rnorm(n),p2=rlnorm(n),p3=runif(n))
# create density plot for each component
gs=densityPlot(sim)
}
