% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getNames}
\alias{getNames}
\title{getNames from an object or a list of objects having a $name field
(e.g. parameters)}
\usage{
getNames(loo, name = "name")
}
\arguments{
\item{loo}{List Of Objects}

\item{name}{character, string denoting the name field}
}
\value{
A character vector containg names
}
\description{
getNames from an object or a list of objects having a $name field
(e.g. parameters)
}
