% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmcSummary.R
\name{mcmcSummary}
\alias{mcmcSummary}
\title{mcmcSummary constructor.}
\usage{
mcmcSummary(fname = "Config_Summary.txt", result.fname = "Results_Summary.txt")
}
\arguments{
\item{fname}{Character, configuration file name.}

\item{result.fname}{Character, result file name.}
}
\value{
An object of class 'mcmcSummary'.
}
\description{
Creates a new instance of a 'mcmcSummary' object
}
\examples{
m <- mcmcSummary()
}
