% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{runExe}
\alias{runExe}
\title{Run BaM executable}
\usage{
runExe(exedir, exename, workspace)
}
\arguments{
\item{exedir}{directory of the executable}

\item{exename}{name of the executable}

\item{workspace}{full path to workspace}
}
\value{
an error code (0 for success)
}
\description{
Run BaM executable
}
\keyword{internal}
