% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{runExe}
\alias{runExe}
\title{Run an executable file}
\usage{
runExe(exedir, exename)
}
\arguments{
\item{exedir}{directory of the executable}

\item{exename}{name of the executable}
}
\value{
Nothing
}
\description{
Run an executable file
}
\keyword{internal}
